requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.3.0

# Base de datos
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36

# Autenticación
Flask-Login==0.6.3

# Validación de formularios
Flask-WTF==1.2.2
WTForms==3.2.1
email-validator==2.3.0

# Email
Flask-Mail==0.10.0
blinker==1.9.0

# Para desarrollo y testing
pytest==8.3.4
pytest-flask==1.3.0

# Para producción
gunicorn==23.0.0
whitenoise==6.8.2

# Para manejo de imágenes
Pillow>=10.3.0

# Utilidades
python-dotenv==1.2.1

# PostgreSQL (para Neon u otros servicios)
psycopg2-binary==2.9.10

# Seguridad y Rate Limiting
Flask-Limiter==4.0.0

# Monitoreo y Error Tracking
sentry-sdk==2.43.0

